Check if file exists before removing the fastlane archive
Locally unlink() seems to work just fine on non-existent files, but on
binary factory that seems to be different for some reason.

+++ toolchain/generate-fastlane-metadata.py
@@ -151,7 +151,8 @@ def downloadScreenshots(applicationName, data):
 def createMetadataArchive(applicationName):
     srcPath = os.path.join(arguments.output, 'metadata')
     zipFileName = os.path.join(srcPath, 'fastlane-' + applicationName + '.zip')
-    os.unlink(zipFileName)
+    if os.path.exists(zipFileName):
+        os.unlink(zipFileName)
     archive = zipfile.ZipFile(zipFileName, 'w')
     archive.write(os.path.join(srcPath, applicationName + '.yml'), applicationName + '.yml')
